+++ src/dpr/components/filter-input/enum.ts
@@ -2,7 +2,7 @@
 export enum FilterType {
   radio = 'Radio',
   dateRange = 'daterange',
-  date = 'Date',
+  date = 'date',
   select = 'Select',
   autocomplete = 'Autocomplete',
   text = 'Text',
